broke out method scanner into its own file
