+++ src/server/frontend_wayland/generator/CMakeLists.txt
@@ -5,6 +5,7 @@ include_directories(SYSTEM ${XMLPP_INCLUDE_DIRS})
 add_executable(wrapper-generator
   wrapper_generator.cpp     wrapper_generator.h
   argument.cpp              argument.h
+  method.cpp                method.h
   emitter.cpp               emitter.h
 )
